# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Notes App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /notes:
    # binds a127 app logic to a route
    x-swagger-router-controller: notes  
    post:
      summary: Post a note 
      operationId: postNote
      parameters:
      - in: body
        name: item
        required: true
        schema:
          $ref: "#/definitions/NoteRequest"
      responses:
        "201":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NoteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   

    put:
      summary: Update a note 
      operationId: putNote
      parameters:
      - in: body
        name: item
        required: true
        schema:
          $ref: "#/definitions/NoteResponse"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   

    get:
      description: Returns 'Notes' to the caller
      # used as the method name of the controller
      operationId: getNotes
      responses:
        "200":
          description: Success
          schema:
            type: array
            # a pointer to a definition
            $ref: "#/definitions/NoteResponse"
          
  /notes/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: notes
    get:
      description: Returns 'Notes' to the caller
      # used as the method name of the controller
      operationId: getNoteById
      parameters:
        - name: id
          in: path
          description: The note of a particular id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NoteResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: Deletes a note
      # used as the method name of the controller
      operationId: deleteNote
      parameters:
        - name: id
          in: path
          description: The note of a particular id
          required: true
          type: string
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  NoteRequest:
    properties:
      title:
        type: string
      desc:
        type: string
  NoteResponse:
    properties:
      title:
        type: string
      desc:
        type: string
      id:
        type: number
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string